#17: Fix items spawning on background and refactor
