+++ Assets/Scripts/DecoratorFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 98301f1478af2850c85241a32b72a2e4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
